+++ src/components/CFilterGroup/CFilterGroup.js
@@ -6,7 +6,6 @@ export default {
   data() {
     return {
       entity: {},
-      errors: [],
     };
   },
   computed: {
