+++ src/components/index.meta.js
@@ -11,7 +11,7 @@ export { default as flexgrid } from './CFlexgrid/index.meta';
 export { default as floatingButton } from './CFloatingButton/index.meta';
 export { default as floatingButtonItem } from './CFloatingButtonItem/index.meta';
 export { default as flexgridItem } from './CFlexgridItem/index.meta';
-export { default as filterForm } from './CFilterForm/index.meta';
+export { default as filterGroup } from './CFilterGroup/index.meta';
 export { default as form } from './CForm/index.meta';
 export { default as gallery } from './CGallery/index.meta';
 export { default as hlist } from './CHlist/index.meta';
